+++ enduro_events/src/components/footer/footer.module.css
@@ -2,7 +2,7 @@
     width: 100%;
     height: 10%;
     background-color: rgba(0,0,0,.5);
-    color: #234465;
+    color: white;
     border-top: 2px solid #234465;
     border-bottom: 2px solid #234465;
     margin-top: 10px;
