+++ enduro_events/src/components/header/header.module.css
@@ -1,61 +1,21 @@
-ul {
-    list-style-type: none;
-    margin: 0;
-    overflow: hidden;
-    margin-top: 6px;
-}
-
-.listItem a {
-    font-family: 'Times New Roman', Times, serif;
-    color: white;
-    font-size: 20px;
-    margin: 2px;
-    padding: 2px;
-    text-decoration: none;
-}
-
 .logo {
-   height: 50px;
-   margin-left: 10px;
-   padding: 3px;
-}
-
-.mountain {
     height: 50px;
-}
+    margin-left: 10px;
+    padding: 3px;
+ }
 
-.bicycle {
-    height: 60px;
-    margin-right: 20px;
-}
-
-.listItem {
-    display: inline-block;
-    height: auto;
-    margin-right: 2%;
-    cursor: pointer;
-    padding: 0.5%;
-    margin: 0.1%;
-    vertical-align: top;
-}
-
-.mainNavigation {
-    padding: 3px 0 0;
-    display: flex;
-    background: rgba(0,0,0,.5);
-    border-bottom: 2px solid #234465;
-}
-
-.navigation {
-    flex: 0 0 35%;
-}
-
-.userNavigation { 
-    flex: 1;
-    text-align: right;
-}
-
-nav .listItem:hover {
-    border-top: none;
-    font-weight: bold;
-}
+ .mountain {
+     height: 50px;
+ }
+ 
+ .bicycle {
+     height: 60px;
+     margin-right: 20px;
+ }
+ 
+ .mainNavigation {
+     padding: 3px 0 0;
+     display: flex;
+     background: rgba(0,0,0,.5);
+     border-bottom: 2px solid #234465;
+ }
